Put the isViaOrthology variable onto the parent so it gets inherited by AllelDiseaseAnnotationDocument
